refactor: move api logic to content component
